stats_bot/requirements.txt
praw
asyncpg
discord.py
matplotlib
buttons
